pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "citeverify"
version = "0.1.0"
description = "Verify citations in research papers from the command line"
authors = [{name = "Your Name", email = "your@email.com"}]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "click>=8.1.0",
    "rich>=13.0.0",
    "pdfplumber>=0.11.0",
    "arxiv>=2.1.0",
    "habanero>=1.2.6",
    "aiohttp>=3.9.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "PyPDF2>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "black>=24.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
citeverify = "citeverify.cli:main"

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
